Extract url functions from netlib.utils and move to netlib.http.url
