+++ netlib/http/http2/connections.py
@@ -6,7 +6,7 @@ import hyperframe.frame
 
 from hpack.hpack import Encoder, Decoder
 from ... import utils
-from .. import Headers, Response, Request
+from .. import Headers, Response, Request, url
 from . import frame
 
 
@@ -118,7 +118,7 @@ class HTTP2Protocol(object):
         else:
             first_line_format = "absolute"
             # FIXME: verify if path or :host contains what we need
-            scheme, host, port, _ = utils.parse_url(path)
+            scheme, host, port, _ = url.parse_url(path)
             scheme = scheme.decode('ascii')
             host = host.decode('ascii')
